Feature/complete todays workout

+++ db/migrate/20220930005259_create_scheduled_workout_exercises.rb
@@ -5,9 +5,9 @@ class CreateScheduledWorkoutExercises < ActiveRecord::Migration[7.0]
     create_table :scheduled_workout_exercises do |t|
       t.references :scheduled_workout, null: false
       t.references :workout_day_exercise, null: false
-      t.numeric :sets, null: false
-      t.numeric :reps, null: false
-      t.numeric :weight, null: false
+      t.integer :sets, null: false
+      t.integer :reps, null: false
+      t.integer :weight, null: false
       t.integer :result, array: true, default: []
       t.timestamps
     end
